Actualización importante: nuevas mejoras al note server Abms regenerados

+++ src/AppBundle/Controller/CompaniaController.php
@@ -2,8 +2,6 @@
 
 namespace AppBundle\Controller;
 
-use Doctrine\DBAL\Driver\PDOException;
-use Symfony\Component\Config\Definition\Exception\Exception;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 
@@ -34,8 +32,11 @@ class CompaniaController extends Controller
             10/* limit per page */
         );
 
+        $deleteForm = $this->createDeleteForm();
+
         return $this->render('AppBundle:compania:index.html.twig', array(
             'companias' => $companias,
+            'delete_form' => $deleteForm->createView()
         ));
     }
 
@@ -55,7 +56,7 @@ class CompaniaController extends Controller
             $em->flush();
 
             // set flash messages
-            $this->get('session')->getFlashBag()->add('success', 'El registro se ha guardaro satisfactoriamente.');
+            $this->get('session')->getFlashBag()->add('success', 'El registro se ha guardado satisfactoriamente.');
 
             return $this->redirectToRoute('compania_index');
 
@@ -73,11 +74,10 @@ class CompaniaController extends Controller
      */
     public function showAction(Compania $companium)
     {
-        $deleteForm = $this->createDeleteForm($companium);
+
 
         return $this->render('AppBundle:compania:show.html.twig', array(
-            'companium' => $companium,
-            'delete_form' => $deleteForm->createView(),
+            'companium' => $companium
         ));
     }
 
@@ -115,21 +115,20 @@ class CompaniaController extends Controller
      */
     public function deleteAction(Request $request, Compania $companium)
     {
+        $form = $this->createDeleteForm($companium);
+        $form->handleRequest($request);
 
-        try{
+        if ($form->isSubmitted() && $form->isValid()) {
+            try{
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($companium);
+                $em->flush();
 
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($companium);
-            $em->flush();
-
-            // set flash messages
-            $this->get('session')->getFlashBag()->add('success', 'La Companía se ha dado de baja satisfactoriamente.');
-
-        }catch (\Exception $ex){
-            $this->get('session')->getFlashBag()->add('error', 'Hubo un error al intentar eliminar la companiía.');
+                $this->get('session')->getFlashBag()->add('success', 'El registro se ha dado de baja satisfactoriamente.');
+            }catch(\Exception $e){
+                $this->get('session')->getFlashBag()->add('error', 'Hubo un error al intentar eliminar el registro.');
+            }
         }
-
-
 
         return $this->redirectToRoute('compania_index');
     }
@@ -137,14 +136,13 @@ class CompaniaController extends Controller
     /**
      * Creates a form to delete a Compania entity.
      *
-     * @param Compania $companium The Compania entity
      *
      * @return \Symfony\Component\Form\Form The form
      */
-    private function createDeleteForm(Compania $companium)
+    private function createDeleteForm()
     {
         return $this->createFormBuilder()
-            ->setAction($this->generateUrl('compania_delete', array('id' => $companium->getId())))
+            ->setAction($this->generateUrl('compania_delete', array('id' => '__obj_id__')))
             ->setMethod('DELETE')
             ->getForm()
         ;
